requirements.txt
appdirs==1.4.3
astroid==2.3.3
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
Django==2.2.3
django-crispy-forms==1.8.0
django-grappelli==2.13.2
future==0.17.1
gmusicapi==12.1.0
gpsoauth==0.4.1
gunicorn==20.0.4
httplib2==0.13.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.3.4
mccabe==0.6.1
MechanicalSoup==0.11.0
mock==3.0.5
mutagen==1.42.0
mysql-connector-python==8.0.16
numpy==1.16.4
oauth2client==4.1.3
pandas==0.24.2
Pillow==6.1.0
proboscis==1.2.6.0
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycryptodomex==3.8.2
pylint==2.4.3
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
validictory==1.1.2
wrapt==1.11.2
wxPython==4.0.6